OCT-111: moved SearchAttributeOptionsQuery
